Initial Sketches
I chose to upload this pdf of my project website sketch into the repo because it allows a more comprehensive way to view my readme and planning process all in one place: my repo. It will also allow me to guide myself while actually working on the html and css for my website.